DEV : refactoring sampling
- grouping base classes in _base
- adding types
