--- src/relife/sampling/__init__.py
@@ -1,9 +1,0 @@
-# """Sampling utilities for lifetime models and stochastic processes."""
-#
-# from ._lifetimes import sample_lifetimes
-# from ._processes import sample_process
-#
-# __all__ = [
-#     "sample_lifetimes",
-#     "sample_process",
-# ]
